requirements.txt
discord.py==1.5.1
apscheduler==3.6.3